Move JsonUtils.kt to the file directory data/parsers
