+++ libsource/src/main/java/com/virtusize/libsource/data/parsers/ProductMetaDataHintsJsonParser.kt
@@ -1,6 +1,5 @@
 package com.virtusize.libsource.data.parsers
 
-import com.virtusize.libsource.data.remote.JsonUtils
 import com.virtusize.libsource.data.remote.ProductMetaDataHints
 import org.json.JSONObject
