+++ libsource/src/main/java/com/virtusize/libsource/network/VirtusizeApiTask.kt
@@ -10,6 +10,7 @@ import com.virtusize.libsource.data.local.message
 import com.virtusize.libsource.data.local.virtusizeError
 import com.virtusize.libsource.data.parsers.VirtusizeJsonParser
 import com.virtusize.libsource.data.remote.ProductCheck
+import com.virtusize.libsource.data.remote.Store
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers.IO
