+++ libsource/src/test/java/com/virtusize/libsource/data/remote/JsonUtilsTest.kt
@@ -1,6 +1,7 @@
 package com.virtusize.libsource.data.remote
 
 import com.google.common.truth.Truth.assertThat
+import com.virtusize.libsource.data.parsers.JsonUtils
 import org.json.JSONArray
 import org.json.JSONObject
 import org.json.JSONObject.NULL
